apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0
pydantic-settings>=2.6.0
pyyaml>=6.0.2
prometheus-client>=0.21.0
slowapi>=0.1.9
PyJWT>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9
scikit-learn>=1.4.0
joblib>=1.3.0
# Dev
pytest>=8.3.0
httpx>=0.28.0
ruff>=0.8.0
